Rename MessageHandler -> EventHandler
